Fix: rename posts folder
