+++ src/pages/posts/[id].tsx
@@ -1,0 +1,14 @@
+import HeadsBase from "~/components/heads";
+
+const SinglePostPage = () => {
+  return (
+    <>
+      <HeadsBase />
+      <div className="p-4 border-b gap-4 border-slate-700 flex items-start">
+        <p>Post view</p>
+      </div>
+    </>
+  );
+}
+
+export default SinglePostPage
